+++ CHANGES.rst
@@ -4,6 +4,15 @@ Changelog
 1.0.9 (unreleased)
 ------------------
 
+- Support showing the original criteria readonly.  This can be handy
+  when you have large selection criteria.  ``show_original_criteria``
+  in the widget is False by default.
+  [maurits]
+
+- Added widget property ``show_parent_criteria``, which is True by
+  default.  This shows the parent criteria readonly in the edit form.
+  [maurits]
+
 - When getting the results, get the ``querybuilderresults`` multi
   adapter, instead of hardcoding the use of the standard
   ``QueryBuilder`` from ``plone.app.querystring``.  This way, add-ons